// HW21/utils/types.ts
export enum SELECTOR_TYPES {
    CLASS_NAME = "className",
    CSS = "css",
    XPATH = "xpath"
};

export enum LANGUAGE {
    RU = 'RU',
    BE = 'BE',
    EN = 'EN'
};

export enum PAGES {
    HOME = "home",
    ADMINPOC = "adminproc",
    APPLICATION = "application",
    ENGLISH = "english",
    SEARCHING = "searching"
};